Renaming to AzureMonitorClient
