+++ sdk/monitor/opentelemetry-exporter-azuremonitor/opentelemetry/exporter/azuremonitor/_generated/__init__.py
@@ -6,8 +6,8 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._application_insights_client import ApplicationInsightsClient
-__all__ = ['ApplicationInsightsClient']
+from ._azure_monitor_client import AzureMonitorClient
+__all__ = ['AzureMonitorClient']
 
 try:
     from ._patch import patch_sdk  # type: ignore
