+++ sdk/monitor/opentelemetry-exporter-azuremonitor/opentelemetry/exporter/azuremonitor/_generated/aio/operations_async/__init__.py
@@ -6,8 +6,8 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._application_insights_client_operations_async import ApplicationInsightsClientOperationsMixin
+from ._azure_monitor_client_operations_async import AzureMonitorClientOperationsMixin
 
 __all__ = [
-    'ApplicationInsightsClientOperationsMixin',
+    'AzureMonitorClientOperationsMixin',
 ]
