+++ sdk/monitor/opentelemetry-exporter-azuremonitor/opentelemetry/exporter/azuremonitor/_generated/models/__init__.py
@@ -43,7 +43,7 @@ except (SyntaxError, ImportError):
     from ._models import TelemetryItem  # type: ignore
     from ._models import TrackResponse  # type: ignore
 
-from ._application_insights_client_enums import (
+from ._azure_monitor_client_enums import (
     ContextTagKeys,
     DataPointType,
     SeverityLevel,
