+++ sdk/monitor/opentelemetry-exporter-azuremonitor/opentelemetry/exporter/azuremonitor/_generated/operations/__init__.py
@@ -6,8 +6,8 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._application_insights_client_operations import ApplicationInsightsClientOperationsMixin
+from ._azure_monitor_client_operations import AzureMonitorClientOperationsMixin
 
 __all__ = [
-    'ApplicationInsightsClientOperationsMixin',
+    'AzureMonitorClientOperationsMixin',
 ]
