+++ sdk/monitor/opentelemetry-exporter-azuremonitor/opentelemetry/exporter/azuremonitor/_generated/operations/_azure_monitor_client_operations.py
@@ -21,7 +21,7 @@ if TYPE_CHECKING:
     T = TypeVar('T')
     ClsType = Optional[Callable[[PipelineResponse[HttpRequest, HttpResponse], T, Dict[str, Any]], Any]]
 
-class ApplicationInsightsClientOperationsMixin(object):
+class AzureMonitorClientOperationsMixin(object):
 
     def track(
         self,
@@ -34,10 +34,10 @@ class ApplicationInsightsClientOperationsMixin(object):
         This operation sends a sequence of telemetry events that will be monitored by Azure Monitor.
 
         :param body: The list of telemetry events to track.
-        :type body: list[~application_insights_client.models.TelemetryItem]
+        :type body: list[~azure_monitor_client.models.TelemetryItem]
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: TrackResponse, or the result of cls(response)
-        :rtype: ~application_insights_client.models.TrackResponse
+        :rtype: ~azure_monitor_client.models.TrackResponse
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["models.TrackResponse"]
